Moved ColorScheme to ColorScheme.cs
